src/deepfake-video/pyproject.toml
[tool.poetry]
name = "deepfake-video"
version = "2.0.0"
description = ""
authors = ["Ben Chou <bchou@umass.edu>", "Matthew Maillet <mcmaillet@umass.edu>", "Om Mehta <omehta@umass.edu>"]

readme = "README.md"
packages = [{include = "deepfake_video"}]
[tool.poetry.scripts]
deepfake-detection = "deepfake_video.main:app"

[tool.poetry.dependencies]
numpy = "*"
pandas = "*"
pillow = "*"
onnxruntime = "*"
dlib = "*"
torchvision = "*"
torch-onnx = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

